feat(issuer): deduplicate attestation proof trust validation
