+++ apps/backend/src/issuer/issuance/oid4vci/deferred-credential.service.ts
@@ -15,7 +15,7 @@ import {
     Openid4vciIssuer,
 } from "@openid4vc/openid4vci";
 import type { Request } from "express";
-import { decodeJwt, decodeProtectedHeader } from "jose";
+import { decodeJwt } from "jose";
 import { Span, TraceService } from "nestjs-otel";
 import { LessThan, Repository } from "typeorm";
 import { v4 } from "uuid";
@@ -23,13 +23,10 @@ import { CryptoService } from "../../../crypto/crypto.service";
 import { Session } from "../../../session/entities/session.entity";
 import { SessionService } from "../../../session/session.service";
 import { TrustStoreService } from "../../../shared/trust/trust-store.service";
-import {
-    ServiceTypeIdentifiers,
-    TrustListSource,
-} from "../../../shared/trust/types";
 import { X509ValidationService } from "../../../shared/trust/x509-validation.service";
 import { CredentialsService } from "../../configuration/credentials/credentials.service";
 import { IssuanceService } from "../../configuration/issuance/issuance.service";
+import { validateAttestationProofTrust } from "./attestation-proof-trust.util";
 import { DeferredCredentialRequestDto } from "./dto/deferred-credential-request.dto";
 import {
     DeferredTransactionEntity,
@@ -93,79 +90,6 @@ export class DeferredCredentialService {
             sessionId,
         );
         return new Openid4vciIssuer({ callbacks });
-    }
-
-    /**
-     * Validate attestation proof signer chain against configured trusted wallet providers.
-     * If no trust list is configured, this check is skipped for backward compatibility.
-     */
-    private async validateAttestationProofTrust(
-        keyAttestationJwt: string,
-        tenantId: string,
-    ): Promise<void> {
-        const issuanceConfig =
-            await this.issuanceService.getIssuanceConfiguration(tenantId);
-        const trustListUrls = issuanceConfig.walletProviderTrustLists ?? [];
-        if (trustListUrls.length === 0) {
-            return;
-        }
-
-        const header = decodeProtectedHeader(keyAttestationJwt);
-        const x5c = header.x5c;
-        if (!Array.isArray(x5c) || x5c.length === 0) {
-            throw new CredentialRequestException(
-                "invalid_proof",
-                "Attestation proof must contain an x5c certificate chain for trust validation",
-            );
-        }
-
-        const trustListSource: TrustListSource = {
-            lotes: trustListUrls.map((url) => ({ url })),
-            acceptedServiceTypes: [ServiceTypeIdentifiers.WalletProvider],
-        };
-
-        const trustStore =
-            await this.trustStoreService.getTrustStore(trustListSource);
-        if (trustStore.entities.length === 0) {
-            throw new CredentialRequestException(
-                "invalid_proof",
-                "No trusted wallet providers found in configured trust lists",
-            );
-        }
-
-        const presentedChain = this.x509ValidationService.parseX5c(x5c);
-        const leaf = presentedChain[0];
-        if (!leaf) {
-            throw new CredentialRequestException(
-                "invalid_proof",
-                "Attestation proof x5c chain is empty",
-            );
-        }
-
-        const anchors = this.x509ValidationService.parseTrustAnchors(
-            trustStore.entities.flatMap((entity) => entity.services),
-        );
-
-        const path = await this.x509ValidationService.buildPath(
-            leaf,
-            presentedChain,
-            anchors,
-        );
-
-        const matched =
-            await this.x509ValidationService.pathMatchesTrustedEntities(
-                path,
-                trustStore.entities,
-                "leaf",
-                ServiceTypeIdentifiers.WalletProvider,
-            );
-
-        if (!matched) {
-            throw new CredentialRequestException(
-                "invalid_proof",
-                "Attestation proof signer is not trusted by configured wallet provider trust lists",
-            );
-        }
     }
 
     /**
@@ -286,7 +210,17 @@ export class DeferredCredentialService {
             });
             holderCnf = verifiedProof.signer.publicJwk as Jwk;
         } else {
-            await this.validateAttestationProofTrust(proof, tenantId);
+            const issuanceConfig =
+                await this.issuanceService.getIssuanceConfiguration(tenantId);
+
+            await validateAttestationProofTrust(
+                proof,
+                issuanceConfig.walletProviderTrustLists ?? [],
+                {
+                    trustStoreService: this.trustStoreService,
+                    x509ValidationService: this.x509ValidationService,
+                },
+            );
 
             const verifiedAttestation =
                 await issuer.verifyCredentialRequestAttestationProof({
